Minor edits

+++ app/assets/javascripts/document/settings/delete.js
@@ -12,19 +12,33 @@ require([
 
     var btnDeleteAnnotations = jQuery('.delete-annotations .btn'),
 
+        deleteAnnotationsEnabled = function() {
+          return !btnDeleteAnnotations.hasClass('disabled');
+        },
+
         deleteAnnotations = function() {
-          var warningTitle = '<span class="icon">&#xf071;</span> Delete All Annotations',
-              warningMsg = '<strong>Are you absolutely sure you want to do this?</strong>',
+          if (deleteAnnotationsEnabled()) {
+            var warningTitle = '<span class="icon">&#xf071;</span> Delete All Annotations',
+                warningMsg = '<strong>Are you absolutely sure you want to do this?</strong>',
 
-              executeDelete = function() {
-                jsRoutes.controllers.document.settings.SettingsController.deleteAnnotations(Config.documentId).ajax()
-                  .done(function(response) {
-                    btnDeleteAnnotations.removeClass('disabled');
-                  });
-              };
+                executeDelete = function() {
+                  jsRoutes.controllers.document.settings.SettingsController.deleteAnnotations(Config.documentId).ajax()
+                    .done(function(response) {
+                      btnDeleteAnnotations.removeClass('disabled');
+                    });
+                },
 
-          btnDeleteAnnotations.addClass('disabled');
-          new Alert('warning', warningTitle, warningMsg).on('ok', executeDelete);
+                resetButton = function() {
+                  btnDeleteAnnotations.removeClass('disabled');
+                },
+
+                alert = new Alert('warning', warningTitle, warningMsg);
+
+            btnDeleteAnnotations.addClass('disabled');
+
+            alert.on('ok', executeDelete);
+            alert.on('cancel', resetButton);
+          }
         };
 
     btnDeleteAnnotations.click(deleteAnnotations);
